Web Project!
The use of the Web (World Wide Web) has had many positive effects on education. It overcomes time and space limitations in traditional schools/colleges. Teachers and students are now using the Web to access vast amounts of information and resources in the cyberspace. Also, learning via the Web enables both synchronous and asynchronous communication. Despite of many benefits of the Web, it may weaken students’ motivation due to lack of face-to-face communication.
The project entitled as “A_Local_Coding_Platform”, which is based on the principles of constructivism, provides motivation and collaborative learning for students in the Web environment. The website consists main features like ‘Coding area’ to provide an IDE for compilation purpose, ‘Problems set’ to sharpen the brain and skills etc. Besides, competition can also be organized on our website locally (without Internet).